requirements.txt
click==8.1.7
fastapi==0.109.0
hydra-core==1.3.2
Markdown==3.5.2
numpy==1.26.3
omegaconf==2.3.0
Pillow==10.2.0
pytest==7.4.4
pytorch_lightning==2.1.3
scikit_learn==1.4.0
starlette==0.35.1
torch==2.1.2
torchmetrics==1.2.1
torchvision==0.16.2
tqdm==4.66.1
uvicorn
pydantic
python-multipart